requirements.txt
pymongo[srv]
python-dotenv
fastapi
uvicorn
